# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrafficAnalysisInput {
  /** @maxLength 2048 */
  url: string;
  /**
     * @minimum 3
     * @maximum 15
     */
  captureDurationSeconds?: number;
  /** Confirms that the user is authorized to analyze this website. */
  authorized: boolean;
}

export type TrafficRequestType = typeof TrafficRequestType[keyof typeof TrafficRequestType];


export const TrafficRequestType = {
  'fetch-xhr': 'fetch-xhr',
  rest: 'rest',
  graphql: 'graphql',
  websocket: 'websocket',
} as const;

/**
 * Value-free JSON Schema containing only type, properties, items and anyOf. Sensitive and opaque field names are omitted.
 * @nullable
 */
export type TrafficRequestJsonSchema = { [key: string]: unknown } | null;

export interface TrafficRequest {
  id: string;
  /** Sanitized origin and path; query strings are always omitted. */
  url: string;
  /** Sanitized hostname of the request. */
  hostname: string;
  /** Sanitized path of the request; query strings are always omitted. */
  path: string;
  method: string;
  /** @nullable */
  statusCode?: number | null;
  /** @nullable */
  contentType?: string | null;
  startedAt: string;
  /** @nullable */
  durationMs?: number | null;
  type: TrafficRequestType;
  resourceType: string;
  /**
     * Value-free JSON Schema containing only type, properties, items and anyOf. Sensitive and opaque field names are omitted.
     * @nullable
     */
  jsonSchema: TrafficRequestJsonSchema;
  isApiCandidate: boolean;
  candidateReasons: string[];
}

export interface TrafficTypeCounts {
  fetchXhr: number;
  rest: number;
  graphql: number;
  websocket: number;
}

export interface TrafficAnalysisResult {
  /** Sanitized analyzed origin and path. */
  url: string;
  capturedAt: string;
  durationMs: number;
  requestCount: number;
  blockedCount: number;
  warnings: string[];
  counts: TrafficTypeCounts;
  requests: TrafficRequest[];
}

export interface TrafficSessionInput {
  /** @maxLength 2048 */
  url: string;
  /** Confirms that the user is authorized to analyze this website. */
  authorized: boolean;
}

export interface TrafficRecordingInput {
  /**
     * @minLength 32
     * @maxLength 64
     */
  sessionId: string;
  recording: boolean;
  /** Clear previous metadata and record only requests initiated from this point. */
  reset?: boolean;
}

export type TrafficInteractionInputAction = typeof TrafficInteractionInputAction[keyof typeof TrafficInteractionInputAction];


export const TrafficInteractionInputAction = {
  click: 'click',
  scroll: 'scroll',
  key: 'key',
} as const;

export interface TrafficInteractionInput {
  /**
     * @minLength 32
     * @maxLength 64
     */
  sessionId: string;
  action: TrafficInteractionInputAction;
  /**
     * @minimum 0
     * @maximum 1280
     */
  x?: number;
  /**
     * @minimum 0
     * @maximum 800
     */
  y?: number;
  /**
     * @minimum -1280
     * @maximum 1280
     */
  deltaX?: number;
  /**
     * @minimum -1600
     * @maximum 1600
     */
  deltaY?: number;
  /**
     * @minLength 1
     * @maxLength 40
     */
  key?: string;
}

export interface TrafficSessionEndInput {
  /**
     * @minLength 32
     * @maxLength 64
     */
  sessionId: string;
}

export interface TrafficSessionState {
  /** @minLength 32 */
  sessionId: string;
  /** Current sanitized origin and path. */
  url: string;
  startedAt: string;
  recording: boolean;
  closed: boolean;
  durationMs: number;
  requestCount: number;
  blockedCount: number;
  warnings: string[];
  counts: TrafficTypeCounts;
  requests: TrafficRequest[];
  /** @nullable */
  screenshot: string | null;
  viewportWidth: number;
  viewportHeight: number;
}

export interface TrafficInteractionResult {
  ok: boolean;
}

export interface ApiError {
  error: string;
}

export type GetTrafficSessionStateParams = {
/**
 * @minLength 32
 * @maxLength 64
 */
sessionId: string;
};