# lib/api-zod/src/generated/types/trafficAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrafficAnalysisInput {
  /** @maxLength 2048 */
  url: string;
  /**
     * @minimum 3
     * @maximum 15
     */
  captureDurationSeconds?: number;
  /** Confirms that the user is authorized to analyze this website. */
  authorized: boolean;
}