# lib/api-zod/src/generated/types/trafficAnalysisResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficRequest } from './trafficRequest';
import type { TrafficTypeCounts } from './trafficTypeCounts';

export interface TrafficAnalysisResult {
  /** Sanitized analyzed origin and path. */
  url: string;
  capturedAt: string;
  durationMs: number;
  requestCount: number;
  blockedCount: number;
  warnings: string[];
  counts: TrafficTypeCounts;
  requests: TrafficRequest[];
}