# lib/api-zod/src/generated/types/trafficInteractionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficInteractionInputAction } from './trafficInteractionInputAction';

export interface TrafficInteractionInput {
  /**
     * @minLength 32
     * @maxLength 64
     */
  sessionId: string;
  action: TrafficInteractionInputAction;
  /**
     * @minimum 0
     * @maximum 1280
     */
  x?: number;
  /**
     * @minimum 0
     * @maximum 800
     */
  y?: number;
  /**
     * @minimum -1280
     * @maximum 1280
     */
  deltaX?: number;
  /**
     * @minimum -1600
     * @maximum 1600
     */
  deltaY?: number;
  /**
     * @minLength 1
     * @maxLength 40
     */
  key?: string;
}