# lib/api-zod/src/generated/types/trafficRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficRequestJsonSchema } from './trafficRequestJsonSchema';
import type { TrafficRequestType } from './trafficRequestType';

export interface TrafficRequest {
  id: string;
  /** Sanitized origin and path; query strings are always omitted. */
  url: string;
  /** Sanitized hostname of the request. */
  hostname: string;
  /** Sanitized path of the request; query strings are always omitted. */
  path: string;
  method: string;
  /** @nullable */
  statusCode?: number | null;
  /** @nullable */
  contentType?: string | null;
  startedAt: string;
  /** @nullable */
  durationMs?: number | null;
  type: TrafficRequestType;
  resourceType: string;
  /**
     * Value-free JSON Schema containing only type, properties, items and anyOf. Sensitive and opaque field names are omitted.
     * @nullable
     */
  jsonSchema: TrafficRequestJsonSchema;
  isApiCandidate: boolean;
  candidateReasons: string[];
}