# lib/api-zod/src/generated/types/trafficSessionEndInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrafficSessionEndInput {
  /**
     * @minLength 32
     * @maxLength 64
     */
  sessionId: string;
}