# lib/api-zod/src/generated/types/trafficSessionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrafficSessionInput {
  /** @maxLength 2048 */
  url: string;
  /** Confirms that the user is authorized to analyze this website. */
  authorized: boolean;
}