# lib/api-zod/src/generated/types/trafficSessionState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficRequest } from './trafficRequest';
import type { TrafficTypeCounts } from './trafficTypeCounts';

export interface TrafficSessionState {
  /** @minLength 32 */
  sessionId: string;
  /** Current sanitized origin and path. */
  url: string;
  startedAt: string;
  recording: boolean;
  closed: boolean;
  durationMs: number;
  requestCount: number;
  blockedCount: number;
  warnings: string[];
  counts: TrafficTypeCounts;
  requests: TrafficRequest[];
  /** @nullable */
  screenshot: string | null;
  viewportWidth: number;
  viewportHeight: number;
}