# lib/api-zod/src/generated/types/trafficTypeCounts.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrafficTypeCounts {
  fetchXhr: number;
  rest: number;
  graphql: number;
  websocket: number;
}